extract iter
