+++ src/rrule.ts
@@ -1,6 +1,4 @@
 import dateutil from './dateutil'
-import Iterinfo, { GetDayset, DaySet } from './iterinfo'
-import { pymod, notEmpty, includes, isPresent } from './helpers'
 
 import IterResult, { IterArgs } from './iterresult'
 import CallbackIterResult from './callbackiterresult'
@@ -8,12 +6,12 @@ import { Language } from './nlp/i18n'
 import { Nlp } from './nlp/index'
 import { GetText } from './nlp/totext'
 import { ParsedOptions, Options, Frequency, QueryMethods } from './types'
-import { parseOptions, initializeOptions, buildTimeset } from './parseoptions'
+import { parseOptions, initializeOptions } from './parseoptions'
 import { parseString } from './parsestring'
 import { optionsToString } from './optionstostring'
 import { Cache, CacheKeys } from './cache'
 import { Weekday } from './weekday'
-import { DateWithZone } from './datewithzone'
+import { iter } from './iter'
 
 interface GetNlp {
   _nlp: Nlp
@@ -134,6 +132,10 @@ export default class RRule implements QueryMethods {
 
   static optionsToString = optionsToString
 
+  protected _iter (iterResult: IterResult) {
+    return iter(iterResult, this.options)
+  }
+
   private _cacheGet (what: CacheKeys | 'all', args?: Partial<IterArgs>) {
     if (!this._cache) return false
     return this._cache._cacheGet(what, args)
@@ -156,11 +158,11 @@ export default class RRule implements QueryMethods {
    */
   all (iterator?: (d: Date, len: number) => boolean): Date[] {
     if (iterator) {
-      return this._iter(new CallbackIterResult('all', {}, iterator), this.options) as Date[]
+      return this._iter(new CallbackIterResult('all', {}, iterator)) as Date[]
     } else {
       let result = this._cacheGet('all') as Date[] | false
       if (result === false) {
-        result = this._iter(new IterResult('all', {}), this.options) as Date[]
+        result = this._iter(new IterResult('all', {})) as Date[]
         this._cacheAdd('all', result)
       }
       return result
@@ -189,14 +191,13 @@ export default class RRule implements QueryMethods {
 
     if (iterator) {
       return this._iter(
-        new CallbackIterResult('between', args, iterator),
-        this.options
+        new CallbackIterResult('between', args, iterator)
       ) as Date[]
     }
 
     let result = this._cacheGet('between', args)
     if (result === false) {
-      result = this._iter(new IterResult('between', args), this.options)!
+      result = this._iter(new IterResult('between', args))!
       this._cacheAdd('between', result, args)
     }
     return result as Date[]
@@ -213,7 +214,7 @@ export default class RRule implements QueryMethods {
     const args = { dt: dt, inc: inc }
     let result = this._cacheGet('before', args)
     if (result === false) {
-      result = this._iter(new IterResult('before', args), this.options)!
+      result = this._iter(new IterResult('before', args))!
       this._cacheAdd('before', result, args)
     }
     return result as Date
@@ -230,7 +231,7 @@ export default class RRule implements QueryMethods {
     const args = { dt: dt, inc: inc }
     let result = this._cacheGet('after', args)
     if (result === false) {
-      result = this._iter(new IterResult('after', args), this.options)!
+      result = this._iter(new IterResult('after', args))!
       this._cacheAdd('after', result, args)
     }
     return result as Date
@@ -272,290 +273,4 @@ export default class RRule implements QueryMethods {
   clone (): RRule {
     return new RRule(this.origOptions)
   }
-
-  _iter (iterResult: IterResult, options: ParsedOptions): Date | Date[] | null {
-    /* Since JavaScript doesn't have the python's yield operator (<1.7),
-        we use the IterResult object that tells us when to stop iterating.
-
-    */
-
-    // Some local variables to speed things up a bit
-    const {
-      dtstart,
-      freq,
-      interval,
-      wkst,
-      until,
-      bymonth,
-      byweekno,
-      byyearday,
-      byweekday,
-      byeaster,
-      bymonthday,
-      bynmonthday,
-      bysetpos,
-      byhour,
-      byminute,
-      bysecond
-    } = options
-
-    let date = new dateutil.DateTime(
-      dtstart.getUTCFullYear(),
-      dtstart.getUTCMonth() + 1,
-      dtstart.getUTCDate(),
-      dtstart.getUTCHours(),
-      dtstart.getUTCMinutes(),
-      dtstart.getUTCSeconds(),
-      dtstart.valueOf() % 1000
-    )
-
-    const ii = new Iterinfo(this)
-    ii.rebuild(date.year, date.month)
-
-    const getdayset = {
-      [RRule.YEARLY]: ii.ydayset,
-      [RRule.MONTHLY]: ii.mdayset,
-      [RRule.WEEKLY]: ii.wdayset,
-      [RRule.DAILY]: ii.ddayset,
-      [RRule.HOURLY]: ii.ddayset,
-      [RRule.MINUTELY]: ii.ddayset,
-      [RRule.SECONDLY]: ii.ddayset
-    }[freq] as GetDayset
-
-    let timeset: dateutil.Time[] | null
-    let gettimeset: () => typeof timeset
-    if (freq < RRule.HOURLY) {
-      timeset = buildTimeset(options)
-    } else {
-      gettimeset = {
-        [RRule.HOURLY]: ii.htimeset,
-        [RRule.MINUTELY]: ii.mtimeset,
-        [RRule.SECONDLY]: ii.stimeset
-      }[
-        freq as Frequency.HOURLY | Frequency.MINUTELY | Frequency.SECONDLY
-      ] as typeof gettimeset
-
-      if (
-        (freq >= RRule.HOURLY &&
-          notEmpty(byhour) &&
-          !includes(byhour, date.hour)) ||
-        (freq >= RRule.MINUTELY &&
-          notEmpty(byminute) &&
-          !includes(byminute, date.minute)) ||
-        (freq >= RRule.SECONDLY &&
-          notEmpty(bysecond) &&
-          !includes(bysecond, date.second))
-      ) {
-        timeset = []
-      } else {
-        timeset = gettimeset.call(
-          ii,
-          date.hour,
-          date.minute,
-          date.second,
-          date.millisecond
-        )
-      }
-    }
-
-    let currentDay: number
-    let count = options.count
-    let pos: number
-
-    while (true) {
-      // Get dayset with the right frequency
-      const [dayset, start, end] = getdayset.call(
-        ii,
-        date.year,
-        date.month,
-        date.day
-      ) as DaySet
-
-      // Do the "hard" work ;-)
-      let filtered = false
-      for (let dayCounter = start; dayCounter < end; dayCounter++) {
-        currentDay = dayset[dayCounter] as number
-
-        filtered = isFiltered(
-          bymonth,
-          ii,
-          currentDay,
-          byweekno,
-          byweekday,
-          byeaster,
-          bymonthday,
-          bynmonthday,
-          byyearday
-        )
-
-        if (filtered) dayset[currentDay] = null
-      }
-
-      // Output results
-      if (notEmpty(bysetpos) && notEmpty(timeset)) {
-        let daypos: number
-        let timepos: number
-        const poslist: Date[] = []
-
-        for (let j = 0; j < bysetpos.length; j++) {
-          pos = bysetpos[j]
-
-          if (pos < 0) {
-            daypos = Math.floor(pos / timeset.length)
-            timepos = pymod(pos, timeset.length)
-          } else {
-            daypos = Math.floor((pos - 1) / timeset.length)
-            timepos = pymod(pos - 1, timeset.length)
-          }
-
-          const tmp = []
-          for (let k = start; k < end; k++) {
-            const val = dayset[k]
-            if (!isPresent(val)) continue
-            tmp.push(val)
-          }
-          let i: number
-          if (daypos < 0) {
-            // we're trying to emulate python's aList[-n]
-            i = tmp.slice(daypos)[0]
-          } else {
-            i = tmp[daypos]
-          }
-
-          const time = timeset[timepos]
-          const date = dateutil.fromOrdinal(ii.yearordinal + i)
-          const res = dateutil.combine(date, time)
-          // XXX: can this ever be in the array?
-          // - compare the actual date instead?
-          if (!includes(poslist, res)) poslist.push(res)
-        }
-
-        dateutil.sort(poslist)
-        for (let j = 0; j < poslist.length; j++) {
-          const res = poslist[j]
-          if (until && res > until) {
-            return emitResult(iterResult)
-          }
-
-          if (res >= dtstart) {
-            const rezonedDate = rezoneIfNeeded(res, options)
-            if (!iterResult.accept(rezonedDate)) {
-              return emitResult(iterResult)
-            }
-
-            if (count) {
-              --count
-              if (!count) {
-                return emitResult(iterResult)
-              }
-            }
-          }
-        }
-      } else {
-        for (let j = start; j < end; j++) {
-          currentDay = dayset[j] as number
-          if (!isPresent(currentDay)) {
-            continue
-          }
-
-          const date = dateutil.fromOrdinal(ii.yearordinal + currentDay)
-          for (let k = 0; k < timeset!.length; k++) {
-            const time = timeset![k]
-            const res = dateutil.combine(date, time)
-            if (until && res > until) {
-              return emitResult(iterResult)
-            }
-
-            if (res >= dtstart) {
-              const rezonedDate = rezoneIfNeeded(res, options)
-              if (!iterResult.accept(rezonedDate)) {
-                return emitResult(iterResult)
-              }
-
-              if (count) {
-                --count
-                if (!count) {
-                  return emitResult(iterResult)
-                }
-              }
-            }
-          }
-        }
-      }
-
-      // Handle frequency and interval
-      if (freq === RRule.YEARLY) {
-        date.addYears(interval)
-      } else if (freq === RRule.MONTHLY) {
-        date.addMonths(interval)
-      } else if (freq === RRule.WEEKLY) {
-        date.addWeekly(interval, wkst)
-      } else if (freq === RRule.DAILY) {
-        date.addDaily(interval)
-      } else if (freq === RRule.HOURLY) {
-        date.addHours(interval, filtered, byhour)
-
-        // @ts-ignore
-        timeset = gettimeset.call(ii, date.hour, date.minute, date.second)
-      } else if (freq === RRule.MINUTELY) {
-        if (date.addMinutes(interval, filtered, byhour, byminute)) {
-          filtered = false
-        }
-
-        // @ts-ignore
-        timeset = gettimeset.call(ii, date.hour, date.minute, date.second)
-      } else if (freq === RRule.SECONDLY) {
-        if (date.addSeconds(interval, filtered, byhour, byminute, bysecond)) {
-          filtered = false
-        }
-
-        // @ts-ignore
-        timeset = gettimeset.call(ii, date.hour, date.minute, date.second)
-      }
-
-      if (date.year > dateutil.MAXYEAR) {
-        return emitResult(iterResult)
-      }
-
-      ii.rebuild(date.year, date.month)
-    }
-  }
-}
-
-function isFiltered (
-  bymonth: number[],
-  ii: Iterinfo,
-  currentDay: number,
-  byweekno: number[],
-  byweekday: number[],
-  byeaster: number | null,
-  bymonthday: number[],
-  bynmonthday: number[],
-  byyearday: number[]
-): boolean {
-  return (
-    (notEmpty(bymonth) && !includes(bymonth, ii.mmask![currentDay])) ||
-    (notEmpty(byweekno) && !ii.wnomask![currentDay]) ||
-    (notEmpty(byweekday) && !includes(byweekday, ii.wdaymask![currentDay])) ||
-    (notEmpty(ii.nwdaymask!) && !ii.nwdaymask![currentDay]) ||
-    (byeaster !== null && !includes(ii.eastermask!, currentDay)) ||
-    ((notEmpty(bymonthday) || notEmpty(bynmonthday)) &&
-      !includes(bymonthday, ii.mdaymask![currentDay]) &&
-      !includes(bynmonthday, ii.nmdaymask![currentDay])) ||
-    (notEmpty(byyearday) &&
-      ((currentDay < ii.yearlen &&
-        !includes(byyearday, currentDay + 1) &&
-        !includes(byyearday, -ii.yearlen + currentDay)) ||
-        (currentDay >= ii.yearlen &&
-          !includes(byyearday, currentDay + 1 - ii.yearlen) &&
-          !includes(byyearday, -ii.nextyearlen + currentDay - ii.yearlen))))
-  )
-}
-
-function rezoneIfNeeded (date: Date, options: ParsedOptions) {
-  return new DateWithZone(date, options.tzid).rezonedDate()
-}
-
-function emitResult (iterResult: IterResult) {
-  return iterResult.getValue() as Date[]
 }
