+++ src/rruleset.ts
@@ -3,6 +3,7 @@ import dateutil from './dateutil'
 import { includes } from './helpers'
 import IterResult from './iterresult'
 import { DateWithZone } from './datewithzone'
+import { iter } from './iter'
 
 export default class RRuleSet extends RRule {
   public readonly _rrule: RRule[]
@@ -42,6 +43,10 @@ export default class RRuleSet extends RRule {
       }
     }
     return undefined
+  }
+
+  _iter (iterResult: IterResult) {
+    return iterSet(iterResult, this._rrule, this._exrule, this._rdate, this._exdate, this.tzid())
   }
 
   /**
@@ -153,73 +158,6 @@ export default class RRuleSet extends RRule {
     return this.valueOf().join('\n')
   }
 
-  _iter (iterResult: IterResult) {
-    const _exdateHash: { [k: number]: boolean } = {}
-    const _exrule = this._exrule
-    const _accept = iterResult.accept
-    const tzid = this.tzid()
-
-    function evalExdate (after: Date, before: Date) {
-      _exrule.forEach(function (rrule) {
-        rrule.between(after, before, true).forEach(function (date) {
-          _exdateHash[Number(date)] = true
-        })
-      })
-    }
-
-    this._exdate.forEach(function (date) {
-      const zonedDate = new DateWithZone(date, tzid).rezonedDate()
-      _exdateHash[Number(zonedDate)] = true
-    })
-
-    iterResult.accept = function (date) {
-      const dt = Number(date)
-      if (!_exdateHash[dt]) {
-        evalExdate(new Date(dt - 1), new Date(dt + 1))
-        if (!_exdateHash[dt]) {
-          _exdateHash[dt] = true
-          return _accept.call(this, date)
-        }
-      }
-      return true
-    }
-
-    if (iterResult.method === 'between') {
-      evalExdate(iterResult.args.after!, iterResult.args.before!)
-      iterResult.accept = function (date) {
-        const dt = Number(date)
-        if (!_exdateHash[dt]) {
-          _exdateHash[dt] = true
-          return _accept.call(this, date)
-        }
-        return true
-      }
-    }
-
-    for (let i = 0; i < this._rdate.length; i++) {
-      const zonedDate = new DateWithZone(this._rdate[i], tzid).rezonedDate()
-      if (!iterResult.accept(new Date(zonedDate.getTime()))) break
-    }
-
-    this._rrule.forEach(function (rrule) {
-      rrule._iter(iterResult, rrule.options)
-    })
-
-    const res = iterResult._result
-    dateutil.sort(res)
-    switch (iterResult.method) {
-      case 'all':
-      case 'between':
-        return res
-      case 'before':
-        return (res.length && res[res.length - 1]) || null
-      case 'after':
-        return (res.length && res[0]) || null
-      default:
-        return null
-    }
-  }
-
   /**
    * Create a new RRuleSet Object completely base on current instance
    */
@@ -239,5 +177,70 @@ export default class RRuleSet extends RRule {
       rrs.exdate(new Date(this._exdate[i].getTime()))
     }
     return rrs
+  }
+}
+
+function iterSet (iterResult: IterResult, _rrule: RRule[], _exrule: RRule[], _rdate: Date[], _exdate: Date[], tzid: string | undefined) {
+  const _exdateHash: { [k: number]: boolean } = {}
+  const _accept = iterResult.accept
+
+  function evalExdate (after: Date, before: Date) {
+    _exrule.forEach(function (rrule) {
+      rrule.between(after, before, true).forEach(function (date) {
+        _exdateHash[Number(date)] = true
+      })
+    })
+  }
+
+  _exdate.forEach(function (date) {
+    const zonedDate = new DateWithZone(date, tzid).rezonedDate()
+    _exdateHash[Number(zonedDate)] = true
+  })
+
+  iterResult.accept = function (date) {
+    const dt = Number(date)
+    if (!_exdateHash[dt]) {
+      evalExdate(new Date(dt - 1), new Date(dt + 1))
+      if (!_exdateHash[dt]) {
+        _exdateHash[dt] = true
+        return _accept.call(this, date)
+      }
+    }
+    return true
+  }
+
+  if (iterResult.method === 'between') {
+    evalExdate(iterResult.args.after!, iterResult.args.before!)
+    iterResult.accept = function (date) {
+      const dt = Number(date)
+      if (!_exdateHash[dt]) {
+        _exdateHash[dt] = true
+        return _accept.call(this, date)
+      }
+      return true
+    }
+  }
+
+  for (let i = 0; i < _rdate.length; i++) {
+    const zonedDate = new DateWithZone(_rdate[i], tzid).rezonedDate()
+    if (!iterResult.accept(new Date(zonedDate.getTime()))) break
+  }
+
+  _rrule.forEach(function (rrule) {
+    iter(iterResult, rrule.options)
+  })
+
+  const res = iterResult._result
+  dateutil.sort(res)
+  switch (iterResult.method) {
+    case 'all':
+    case 'between':
+      return res
+    case 'before':
+      return (res.length && res[res.length - 1]) || null
+    case 'after':
+      return (res.length && res[0]) || null
+    default:
+      return null
   }
 }
